The 'removeFile' function in 'fileutil.py' had a check added to verify whether or not a GEIS data file existed to be removed. This allowed non-image files to be passed to this routine for deletion without causing an error. WJH
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/pydrizzle@194 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/fileutil.py
@@ -659,7 +659,11 @@ def _remove(file):
         # Assumption: filenames end in '.??h' and '.??d'
         #
         os.remove(file)
-        os.remove(file[:-1]+'d')
+        # At this point, we may be deleting a non-image
+        # file, so only verify whether a GEIS hhd or similar
+        # file exists before trying to delete it.
+        if findFile(file[:-1]+'d'):
+            os.remove(file[:-1]+'d')
 
 def removeFile(inlist):
     """ Utility function for deleting a list of files or a single file. """
